# contratos/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "auto_cusco API",
    "description": "API REST del backend de auto_cusco: ingesta de sabanas, seleccion y segmentacion de cartera, generacion de cargas (digitales y VoIP) y reportes de cobranza.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/cartera": {
      "get": {
        "tags": [
          "cartera"
        ],
        "summary": "Consultar Cartera",
        "description": "Productos de la version vigente de la fecha (RF-04, RF-05, RF-08).",
        "operationId": "consultar_cartera_cartera_get",
        "parameters": [
          {
            "name": "fecha_corte",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Fecha Corte"
            }
          },
          {
            "name": "filtro",
            "in": "query",
            "required": false,
            "schema": {
              "type": "array",
              "items": {
                "type": "string"
              },
              "default": [],
              "title": "Filtro"
            }
          },
          {
            "name": "orden",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "campo, o -campo para descendente",
              "title": "Orden"
            },
            "description": "campo, o -campo para descendente"
          },
          {
            "name": "limite",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 5000,
              "minimum": 1,
              "default": 50,
              "title": "Limite"
            }
          },
          {
            "name": "desplazamiento",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Desplazamiento"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginaRespuesta"
                }
              }
            }
          },
          "400": {
            "description": "La peticion no se puede atender tal como viene; el motivo esta en detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/selecciones": {
      "get": {
        "tags": [
          "selecciones"
        ],
        "summary": "Listar Selecciones",
        "description": "Todas las selecciones, ordenadas por nombre y revisadas contra el catalogo actual.",
        "operationId": "listar_selecciones_selecciones_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SeleccionRespuesta"
                  },
                  "type": "array",
                  "title": "Response Listar Selecciones Selecciones Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "selecciones"
        ],
        "summary": "Crear Seleccion",
        "description": "Guarda una seleccion nueva. Si alguna parte no aplica, se rechaza con el motivo.",
        "operationId": "crear_seleccion_selecciones_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeleccionEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeleccionRespuesta"
                }
              }
            }
          },
          "400": {
            "description": "La peticion no se puede atender tal como viene; el motivo esta en detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "409": {
            "description": "La operacion choca con el estado actual",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/selecciones/revision": {
      "post": {
        "tags": [
          "selecciones"
        ],
        "summary": "Revisar Seleccion",
        "description": "Que partes no aplican, sin guardar nada: para marcar errores mientras se edita.",
        "operationId": "revisar_seleccion_selecciones_revision_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeleccionEntrada"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RevisionRespuesta"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/selecciones/{seleccion_id}": {
      "get": {
        "tags": [
          "selecciones"
        ],
        "summary": "Obtener Seleccion",
        "description": "Una seleccion. Si perdio validez se devuelve igual, con sus problemas marcados.",
        "operationId": "obtener_seleccion_selecciones__seleccion_id__get",
        "parameters": [
          {
            "name": "seleccion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Seleccion Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeleccionRespuesta"
                }
              }
            }
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "selecciones"
        ],
        "summary": "Actualizar Seleccion",
        "description": "Reemplaza la seleccion completa.",
        "operationId": "actualizar_seleccion_selecciones__seleccion_id__put",
        "parameters": [
          {
            "name": "seleccion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Seleccion Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SeleccionEntrada"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SeleccionRespuesta"
                }
              }
            }
          },
          "400": {
            "description": "La peticion no se puede atender tal como viene; el motivo esta en detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "409": {
            "description": "La operacion choca con el estado actual",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "selecciones"
        ],
        "summary": "Eliminar Seleccion",
        "operationId": "eliminar_seleccion_selecciones__seleccion_id__delete",
        "parameters": [
          {
            "name": "seleccion_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Seleccion Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "404": {
            "description": "No existe lo pedido",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleError"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "DetalleError": {
        "properties": {
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "detail"
        ],
        "title": "DetalleError",
        "description": "Motivo legible por el que la API rechazo la peticion."
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PaginaRespuesta": {
        "properties": {
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limite": {
            "type": "integer",
            "title": "Limite"
          },
          "desplazamiento": {
            "type": "integer",
            "title": "Desplazamiento"
          },
          "suficiente": {
            "type": "boolean",
            "title": "Suficiente"
          },
          "productos": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Productos"
          }
        },
        "type": "object",
        "required": [
          "total",
          "limite",
          "desplazamiento",
          "suficiente",
          "productos"
        ],
        "title": "PaginaRespuesta"
      },
      "ParteSeleccion": {
        "type": "string",
        "enum": [
          "nombre",
          "filtro",
          "orden",
          "cantidad",
          "indicador"
        ],
        "title": "ParteSeleccion"
      },
...
      "ProblemaRespuesta": {
        "properties": {
          "parte": {
            "$ref": "#/components/schemas/ParteSeleccion"
          },
          "expresion": {
            "type": "string",
            "title": "Expresion"
          },
          "detalle": {
            "type": "string",
            "title": "Detalle"
          }
        },
        "type": "object",
        "required": [
          "parte",
          "expresion",
          "detalle"
        ],
        "title": "ProblemaRespuesta"
      },
...
      "RevisionRespuesta": {
        "properties": {
          "aplicable": {
            "type": "boolean",
            "title": "Aplicable"
          },
          "problemas": {
            "items": {
              "$ref": "#/components/schemas/ProblemaRespuesta"
            },
            "type": "array",
            "title": "Problemas"
          }
        },
        "type": "object",
        "required": [
          "aplicable",
          "problemas"
        ],
        "title": "RevisionRespuesta"
      },
...
      "SeleccionEntrada": {
        "properties": {
          "nombre": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Nombre"
          },
          "filtros": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Filtros",
            "default": []
          },
          "orden": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orden"
          },
          "cantidad": {
            "anyOf": [
              {
                "type": "integer",
                "maximum": 120000.0,
                "minimum": 1.0
              },
              {
                "type": "null"
              }
            ],
            "title": "Cantidad"
          },
          "indicadores": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Indicadores",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "nombre"
        ],
        "title": "SeleccionEntrada"
      },
      "SeleccionRespuesta": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "nombre": {
            "type": "string",
            "title": "Nombre"
          },
          "filtros": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Filtros"
          },
          "orden": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Orden"
          },
          "cantidad": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cantidad"
          },
          "indicadores": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Indicadores"
          },
          "creado_en": {
            "type": "string",
            "format": "date-time",
            "title": "Creado En"
          },
          "actualizado_en": {
            "type": "string",
            "format": "date-time",
            "title": "Actualizado En"
          },
          "aplicable": {
            "type": "boolean",
            "title": "Aplicable"
          },
          "problemas": {
            "items": {
              "$ref": "#/components/schemas/ProblemaRespuesta"
            },
            "type": "array",
            "title": "Problemas"
          }
        },
        "type": "object",
        "required": [
          "id",
          "nombre",
          "filtros",
          "orden",
          "cantidad",
          "indicadores",
          "creado_en",
          "actualizado_en",
          "aplicable",
          "problemas"
        ],
        "title": "SeleccionRespuesta"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}